requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.9
cycler==0.10.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isodate==0.6.1
kiwisolver==1.3.1
lxml==4.7.1
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycountry==20.7.3
pycryptodome==3.12.0
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
streamlink==3.0.3
toml==0.10.2
twilio==6.51.1
uritemplate==3.0.1
urllib3==1.26.2
websocket-client==1.2.3
